V2.6.0
Made the code even cleaner! Fixed the bug where in the custom sort popup sometimes the list it out of sync with the chart. Also fixed the cancel button. In addition I gave the ability to name the Total series whatever you want in a widget script. Still working on the styling of the custom sort page.